1.add/remove comments, remove unused code, move fn to helpers
