+++ index.js
@@ -1,6 +1,9 @@
+//------------------------------------------------------//
 const port = process.env.PORT || 3000
 const dbUrl = process.env.NODE_ENV === 'production' ?
   process.env.MONGODB_URI : 'mongodb://localhost/project2'
+
+//------------------------------------------------------//
 
 const request = require('request-promise-native')
 const express = require('express')
@@ -10,7 +13,7 @@ const {
   isLoggedIn
 } = require('./helpers')
 const User = require('./models/user')
-const Stop = require('./models/stop') //to chnage to route later
+const Stop = require('./models/stop') //to change to route later
 const BusService = require('./models/busService')
 const path = require('path') // for Public files
 const exphbs = require('express-handlebars') // for Handlebars
@@ -20,6 +23,8 @@ const session = require('express-session')
 const MongoStore = require('connect-mongo')(session)
 const passport = require('./config/ppConfig');
 
+//------------------------------------------------------//
+
 require('dotenv').config({
   silent: true
 })
@@ -27,16 +32,10 @@ require('dotenv').config({
 const tptApiKey = process.env.APIKEY
 const geoMapKey = process.env.GEO
 
-
-const home_routes = require('./routes/home_routes')
-const register_routes = require('./routes/register_routes')
-const login_routes = require('./routes/login_routes')
-const load_routes = require('./routes/load_routes')
-const stop_routes = require('./routes/stop_routes')
-// const save_routes = require('./routes/save_routes')
+//------------------------------------------------------//
 
 const app = express()
-// VIEW ENGINES aka handlebars setup
+
 app.engine('handlebars', exphbs({
   defaultLayout: 'main'
 }))
@@ -53,6 +52,8 @@ app.use(bodyParser.urlencoded({
 }))
 app.use(methodOverride('_method'))
 
+//------------------------------------------------------//
+
 mongoose.Promise = global.Promise
 mongoose.connect(dbUrl, {
     useMongoClient: true
@@ -65,6 +66,8 @@ mongoose.connect(dbUrl, {
       console.log(err)
     }
   )
+
+//------------------------------------------------------//
 //session
 app.use(session({
   secret: process.env.SESSION_SECRET,
@@ -77,70 +80,24 @@ app.use(session({
 app.use(passport.initialize());
 app.use(passport.session());
 
+//------------------------------------------------------//
+
 app.use((req, res, next) => {
   app.locals.user = req.user // we'll only `req.user` if we managed to log in
   // console.log(req.user.id);
   next() //make sure other requests continue!
 })
 
-app.get('/save/stops/:code', (req, res) => {
-//shd this be a post request instead?
-  //check if user is logged in. if not, send to login page.
-  if(!req.user){
-      res.redirect('/login')
-  }
-  else{
-    var busStop = req.params.code
-    User.findByIdAndUpdate(req.user.id, {
-      $addToSet: {
-        faveStops: busStop
-      }
-    })
-    .then(console.log('found user.'))
-    .catch(err => console.log(err))
-    // .then(res.send(user))
-  }
-})
+//------------------------------------------------------//
+//Set up routes here
+const home_routes = require('./routes/home_routes')
+const register_routes = require('./routes/register_routes')
+const login_routes = require('./routes/login_routes')
+const load_routes = require('./routes/load_routes')
+const stop_routes = require('./routes/stop_routes')
 
-app.get('/save/bus/:code', (req, res) => {
-  if(!req.user){
-      res.redirect('/login')
-  }
-  // var bus = req.body.busNo
-  else{
-    var bus = req.params.code
-    User.findByIdAndUpdate(req.user.id, {
-      $addToSet: {
-        faveBus: bus
-      }
-    })
-    .then(console.log('found user.'))
-      .catch(err => console.log(err))
+//------------------------------------------------------//
 
-      // .then(() => {
-      //   res.send({
-      //     message: 'added with success'
-      //   })
-        // }
-    // .then(res.send(user))
-
-  }//end else
-})//end get save
-
-
-
-
-app.get('/profile', (req, res) => {
-  // User.findOne({
-  //     slug: req.params.slug
-  //   })
-  //   .then((user) => {
-  //     res.render('users/show', {
-  //       user
-  //     })
-  //   })
-  res.send(req.user)
-})
 app.use('/', home_routes)
 app.use('/stop', stop_routes)
 // app.use('/save', save_routes)
@@ -151,6 +108,50 @@ app.get('/logout', hasLoggedOut, (req, res) => {
   req.logout()
   res.redirect('/')
 })
+
+//--------------------Simple routes using req.user.id here-----------------------//
+//when saving bus stops
+//Phase 2 - todo add more functionality with saved bus stops
+//show on dashboard(home) even if not nearby (reference for route planning)
+app.get('/save/stops/:code', (req, res) => {
+  //shd this be a post request instead?
+  //check if user is logged in. if not, send to login page.
+  if (!req.user) {
+    res.redirect('/login')
+  } else {
+    var busStop = req.params.code
+    User.findByIdAndUpdate(req.user.id, {
+        $addToSet: {
+          faveStops: busStop
+        }
+      })
+      .catch(err => console.log(err))
+    // .then(res.send(user))
+  }
+})
+//when saving buses
+//Phase 2 - todo - add more functionality with saved bus stops
+//show saved buses on dashboard(home) if stop is available nearby
+app.get('/save/bus/:code', (req, res) => {
+  if (!req.user) {
+    res.redirect('/login')
+  }
+  // var bus = req.body.busNo
+  else {
+    var bus = req.params.code
+    User.findByIdAndUpdate(req.user.id, {
+        $addToSet: {
+          faveBus: bus
+        }
+      })
+      .then(console.log('found user.'))
+      .catch(err => console.log(err))
+  } //end else
+}) //end get save
+
+//removed user profile - not needed for app
+
+//------------------------------------------------------//
 
 app.listen(port, () => {
   console.log(`Server is running on ${port}`)
