+++ routes/login_routes.js
@@ -5,7 +5,6 @@ const router = express.Router()
 const passport = require('../config/ppConfig');
 
 router.get('/', (req, res) => {
-//when rendering, your POV is under `views`
   res.render('users/login')
 })
