+++ routes/stop_routes.js
@@ -29,20 +29,18 @@ router.get('/:code', (req, res) => {
         //convert string to usable format
         var sub = str.substring(str.indexOf('T') + 1, str.indexOf('+'))
         var sub2 = str2.substring(str2.indexOf('T') + 1, str2.indexOf('+'))
-        console.log(sub)
-        console.log(sub2)
         //using moment package, get relative time that shows in arrival time in natural language
         var diff = moment(sub, "h:mm:ss").fromNow()
         var diff2 = moment(sub2, "h:mm:ss").fromNow()
-        console.log(diff);
-        console.log(diff2);
         var num = arr[i].ServiceNo
-        console.log(num);
         buses.push({
-          num, diff,diff2
+          num,
+          diff,
+          diff2
         })
       }
       console.log(buses)
+      console.log('in stop routes')
       res.render('stop', {
         buses: buses
       }) //send only service and time diff
