+++ views/home.handlebars
@@ -9,3 +9,13 @@
 <div class="map">
   <img src="" alt="" class="gmap">
 </div>
+
+{{!-- todo Phase 2 only
+<div class="saved">
+{{#each stops as |stop|}}
+{{stop.stopCode}}
+{{stop.description}}
+{{stop.road}}
+{{/each}}
+
+</div> --}}
