--- views/partials/stopspartial.handlebars
@@ -1,10 +1,0 @@
-<div class="">
-  <div>
-    testing
-  </div>
-  {{!-- {{nearest[0].stopCode}} --}}
-  {{#each nearest as |stop|}}
-    {{stop.stopCode}}
-  {{/each}}
-
-</div>
